+++ c-api/entities/colshape.cpp
@@ -13,7 +13,7 @@ uint8_t ColShape_IsEntityIn(alt::IColShape* colShape, alt::IEntity* entity) {
     return colShape->IsEntityIn(entity);
 }
 
-uint8_t ColShape_IsEntityIdIn(alt::IColShape* colShape, uint16_t id) {
+uint8_t ColShape_IsEntityIdIn(alt::IColShape* colShape, uint32_t id) {
     return colShape->IsEntityIdIn(id);
 }
