+++ c-api/entities/colshape.h
@@ -19,7 +19,7 @@ EXPORT_SHARED uint32_t ColShape_GetID(alt::IColShape* colShape);
 EXPORT_SHARED alt::IWorldObject* ColShape_GetWorldObject(alt::IColShape* colShape);
 
 EXPORT_SHARED uint8_t ColShape_IsEntityIn(alt::IColShape* colShape, alt::IEntity* entity);
-EXPORT_SHARED uint8_t ColShape_IsEntityIdIn(alt::IColShape* colShape, uint16_t id);
+EXPORT_SHARED uint8_t ColShape_IsEntityIdIn(alt::IColShape* colShape, uint32_t id);
 EXPORT_SHARED uint8_t ColShape_IsPointIn(alt::IColShape* colShape, vector3_t point);
 
 EXPORT_SERVER uint8_t ColShape_GetColShapeType(alt::IColShape* colShape);
